Refactor reset into its own file
